Resolved CID 351528 (#1 of 1): Invalid type in argument to printf format specifier
invalid_type: Argument this->mTrainers.size() to format specifier %u was expected to have type unsigned int but has type unsigned long

+++ src/world/Objects/ObjectMgr.cpp
@@ -1688,7 +1688,7 @@ void ObjectMgr::LoadTrainers()
 
     } while (result->NextRow());
     delete result;
-    LogDetail("ObjectMgr : %u trainers loaded.", mTrainers.size());
+    LogDetail("ObjectMgr : %u trainers loaded.", static_cast<uint32_t>(mTrainers.size()));
 }
 #else
 void ObjectMgr::LoadTrainers()
@@ -1848,7 +1848,7 @@ void ObjectMgr::LoadTrainers()
     }
     while (result->NextRow());
     delete result;
-    LogDetail("ObjectMgr : %u trainers loaded.", mTrainers.size());
+    LogDetail("ObjectMgr : %u trainers loaded.", static_cast<uint32_t>(mTrainers.size()));
 #endif
 }
 #endif
